+++ src/mol-repr/volume/slice.ts
@@ -35,6 +35,9 @@ export const SliceParams = {
         x: PD.Numeric(0, { min: 0, max: 0, step: 1 }, { immediateUpdate: true }),
         y: PD.Numeric(0, { min: 0, max: 0, step: 1 }, { immediateUpdate: true }),
         z: PD.Numeric(0, { min: 0, max: 0, step: 1 }, { immediateUpdate: true }),
+        relativeX: PD.Numeric(0, { min: 0, max: 1, step: 0.01 }, { immediateUpdate: true }),
+        relativeY: PD.Numeric(0, { min: 0, max: 1, step: 0.01 }, { immediateUpdate: true }),
+        relativeZ: PD.Numeric(0, { min: 0, max: 1, step: 0.01 }, { immediateUpdate: true }),
     }, { isEssential: true, hideIf: p => p.mode !== 'grid', description: 'Slice position in grid coordinates.' }),
     isoValue: Volume.IsoValueParam,
     mode: PD.Select('grid', PD.arrayToOptions(['grid', 'frame', 'plane'] as const), { description: 'Grid: slice through the volume along the grid axes in integer steps. Frame: slice through the volume along arbitrary axes in any step size. Plane: an arbitrary plane defined by point and normal.' }),
@@ -59,6 +62,9 @@ export function getSliceParams(ctx: ThemeRegistryContext, volume: Volume) {
         x: PD.Numeric(0, { min: 0, max: dim[0] - 1, step: 1 }, { immediateUpdate: true }),
         y: PD.Numeric(0, { min: 0, max: dim[1] - 1, step: 1 }, { immediateUpdate: true }),
         z: PD.Numeric(0, { min: 0, max: dim[2] - 1, step: 1 }, { immediateUpdate: true }),
+        relativeX: PD.Numeric(0, { min: 0, max: 1, step: 0.01 }, { immediateUpdate: true }),
+        relativeY: PD.Numeric(0, { min: 0, max: 1, step: 0.01 }, { immediateUpdate: true }),
+        relativeZ: PD.Numeric(0, { min: 0, max: 1, step: 0.01 }, { immediateUpdate: true }),
     }, { isEssential: true, hideIf: p => p.mode !== 'grid' });
     p.isoValue = Volume.createIsoValueParam(Volume.IsoValue.absolute(volume.grid.stats.min), volume.grid.stats);
     return p;
@@ -329,8 +335,8 @@ async function createGridImage(ctx: VisualContext, volume: Volume, key: number, 
     } = getSliceInfo(volume.grid, props);
 
     const corners = new Float32Array(
-        dim === 'x' ? [x, 0, 0, x, y, 0, x, 0, z, x, y, z] :
-            dim === 'y' ? [0, y, 0, x, y, 0, 0, y, z, x, y, z] :
+        dim === 'x' || dim === 'relativeX' ? [x, 0, 0, x, y, 0, x, 0, z, x, y, z] :
+            dim === 'y' || dim === 'relativeY' ? [0, y, 0, x, y, 0, 0, y, z, x, y, z] :
                 [0, 0, z, 0, y, z, x, 0, z, x, y, z]
     );
 
@@ -379,6 +385,10 @@ async function createGridImage(ctx: VisualContext, volume: Volume, key: number, 
 
 //
 
+function getRelativeIndex(dim: number, index: number) {
+    return clamp(Math.round((dim - 1) * index), 0, dim - 1);
+}
+
 function getSliceInfo(grid: Grid, props: SliceProps) {
     const { dimension: { name: dim, params: index } } = props;
     const { space } = grid.cells;
@@ -396,8 +406,26 @@ function getSliceInfo(grid: Grid, props: SliceProps) {
         x = nx - 1, y = index, z = nz - 1;
         width = nz, height = nx;
         y0 = y, ny = y0 + 1;
-    } else {
+    } else if (dim === 'z') {
         x = nx - 1, y = ny - 1, z = index;
+        width = nx, height = ny;
+        z0 = z, nz = z0 + 1;
+    } else if (dim === 'relativeX') {
+        x = getRelativeIndex(nx, index);
+        y = ny - 1;
+        z = nz - 1;
+        width = nz, height = ny;
+        x0 = x, nx = x0 + 1;
+    } else if (dim === 'relativeY') {
+        x = nx - 1;
+        y = getRelativeIndex(ny, index);
+        z = nz - 1;
+        width = nz, height = nx;
+        y0 = y, ny = y0 + 1;
+    } else /* if (dim === 'relativeZ') */ {
+        x = nx - 1;
+        y = ny - 1;
+        z = getRelativeIndex(nz, index);
         width = nx, height = ny;
         z0 = z, nz = z0 + 1;
     }